+++ marko/__init__.py
@@ -14,6 +14,7 @@ from .html_renderer import HTMLRenderer
 from .renderer import Renderer
 from .parser import Parser
 from .helpers import is_type_check, load_extension_object
+from ._compat import string_types
 
 if is_type_check():
     from typing import Type, List, Any, Optional
@@ -72,7 +73,7 @@ class Markdown(object):
         if self._setup_done:
             raise SetupDone()
         for extension in extensions:
-            if isinstance(extension, str):
+            if isinstance(extension, string_types):
                 extension = load_extension_object(extension)
 
             self._parser_mixins = (